requirements.txt
django~=5.0.3
psycopg2-binary
pillow
djangorestframework~=3.14.0
python-dotenv~=1.0.1

django-filter~=24.1